move tag to its own projection
