+++ main-site/lib/groq/tag.projection.ts
@@ -1,0 +1,23 @@
+import groq from 'groq';
+
+export type Tag = {
+  isLowConfidence: boolean | null;
+  isBackground: boolean | null;
+  tag: {
+    name: string;
+    issue: {
+      name: string;
+    };
+  };
+};
+
+export const tagProjection = groq`
+tags[]{
+  _id,
+  isLowConfidence,
+  isBackground,
+  tag->{
+    name,
+    'issue': topic->{name}
+  }
+}`;
